+++ ImageProcessorSources/main/kotlin/IFilter.kt
@@ -1,5 +1,5 @@
 import java.awt.image.BufferedImage
 
 interface IFilter {
-    fun apply(image: BufferedImage, filterStrength: Int, additionalData: List<Any> = emptyList()) : BufferedImage
+    fun apply(image: BufferedImage, filterStrength: Float, additionalData: List<Any> = emptyList()) : BufferedImage
 }
